requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
SQLAlchemy==2.0.36
openai==1.44.0
python-dotenv==1.0.1
